fixed image orientation in handwritten invoive images for easier extraction

+++ main.py
@@ -1,10 +1,10 @@
-from fastapi import FastAPI, UploadFile, File
+from fastapi import FastAPI, UploadFile, File, Form
 from dotenv import load_dotenv
 from database import SessionLocal
 from models import Invoice
 from validation import validate_invoice
 from ocr import extract_text_from_image
-from extraction import extract_from_printed_text
+from extraction import extract_from_printed_text, extract_from_handwritten_image
 import shutil
 import os
 
@@ -21,41 +21,43 @@ def health():
     return {"status": "healthy"}
 
 @app.post("/upload-invoice")
-async def upload_invoice(file: UploadFile = File(...)):
+async def upload_invoice(file: UploadFile = File(...),doc_type: str = Form(...)):
     # temp save the uploaded file
     temp_path = f"temp_{file.filename}"
     with open(temp_path, "wb") as buffer:
         shutil.copyfileobj(file.file, buffer)
 
     # OCR
-    raw_text = extract_text_from_image(temp_path)
+    try:
+        if doc_type == "handwritten":
+            raw_text = None
+            extracted_data = extract_from_handwritten_image(temp_path)
+        else:
+            raw_text = extract_text_from_image(temp_path)
+            extracted_data = extract_from_printed_text(raw_text)
 
-    # Gemini extraction
-    extracted_data = extract_from_printed_text(raw_text)
-    # save to DB
-    
-    db = SessionLocal()
+        db = SessionLocal()
+        status, reason = validate_invoice(extracted_data, db, Invoice)
 
-    status, reason = validate_invoice(extracted_data, db, Invoice)
-    
-    invoice = Invoice(
-        filename=file.filename,
-        raw_ocr_text=raw_text,
-        extracted_data=extracted_data,
-        validation_status=status,
-        validation_reason=reason
-    )
-    db.add(invoice)
-    db.commit()
-    db.refresh(invoice)
-    db.close()
+        invoice = Invoice(
+            filename=file.filename,
+            raw_ocr_text=raw_text,
+            extracted_data=extracted_data,
+            validation_status=status,
+            validation_reason=reason
+        )
+        db.add(invoice)
+        db.commit()
+        db.refresh(invoice)
+        db.close()
 
-    os.remove(temp_path)
-
-    return {
-        "invoice_id": invoice.id,
-        "extracted_data": extracted_data,
-        "validation_status": status,
-        "validation_reason": reason
-
-    }
+        return {
+            "invoice_id": invoice.id,
+            "doc_type": doc_type,
+            "extracted_data": extracted_data,
+            "validation_status": status,
+            "validation_reason": reason
+        }
+    finally:
+        if os.path.exists(temp_path):
+            os.remove(temp_path)
